fix(perf+wgsl): named GL enums + WGSL polygon mod + browser-support hint
TimerQueryPool: capture enum values from the extension object at runtime
(ext.TIME_ELAPSED_EXT etc.) with the registry hex literals as documented
fallback. Removes magic-number ambiguity; verbose enums now have names
that match the spec. Hardware extension polyfills that don't expose the
named properties still work via the fallback path.

Constructor now console.warn's when the timer-query extension is
unavailable (Firefox without the privileged-extensions flag, very old
Safari, headless software-rendered Chrome on some configs) so users
understand why getGpuTimings() returns null instead of staring at an
empty benchmark page.

find-points-in-polygon.wgsl: replaced float `%` ring-index math with
i32 `%`. The float form worked today because operands are non-negative
but WGSL's float `%` is IEEE remainder (not GLSL floor-mod) and would
silently break if anyone refactored to allow negative path indices.

+++ src/modules/Points/find-points-in-polygon.wgsl
@@ -58,7 +58,10 @@ fn pointInPolygon(point: vec2f, pathLength: i32) -> bool {
       break;
     }
 
-    let j = i32(((f32(i) + 1.0) % f32(pathLength)));
+    // (i + 1) % pathLength on i32 — well-defined for non-negative ints,
+    // and safe under any future refactor that might introduce negatives
+    // (unlike float `%` which is IEEE remainder, not floor mod).
+    let j = (i + 1) % pathLength;
 
     let pi = getPolygonPoint(i, pathLength);
     let pj = getPolygonPoint(j, pathLength);
